changes to work with stock arduino 1.6.9 and above

+++ cores/tivac/WMath.cpp
@@ -28,7 +28,7 @@ extern "C" {
 /* Using interal random and srandom in file random.c
  * until msp430-libc adds supports for random and srandom */
 extern long random(void);
-extern void srandom(unsigned long __seed);
+extern void srandom(unsigned int __seed);
 }
 
 void randomSeed(unsigned int seed)
